Fix alignment when merging zips.
When merging a zip into an apk, apkzlib used to copy all
the non-ignored files from the zip as-is. This was done to
improve performance as it avoid recompressing files that
were already compressed.

However, if the files need to be uncompressed (for example,
native libraries in M+ devices), this would not ensure that
the libraries were uncompressed: if the libraries were
compressed in the original zip file, they were copied
as-is, i.e., compressed, to the apk breaking alignment.

This CL fixes this by ensuring that files that *need* to
be uncompressed (and only those) are actually added to
the apk following compression and alignment rules.

Test: included
Bug: http://b/37926537
Change-Id: I69848b37ef33b4f0af07dde8c778c7823443d55b